+++ examples/react-teleport/src/Teleport.tsx
@@ -20,16 +20,14 @@ const teleportApis = {
 
 const TxStatus: React.FC<{ status: TxEvent | null }> = ({ status }) => {
   if (!status) return null
-  if (status.type === "signed") return <div>Tx Signed {status.txHash}</div>
-  if (status.type === "broadcasted")
+  if (status.type === "created") return <div>Tx Signed {status.txHash}</div>
+  if (status.type === "broadcasted" || status.type === "notInBestBlock")
     return <div>Tx Broadcasted {status.txHash}</div>
-  if (status.type === "txBestBlocksState")
-    return status.found ? (
+  if (status.type === "inBestBlock")
+    return (
       <div>
         Tx included in best block {status.block.hash}-{status.block.index}
       </div>
-    ) : (
-      <div>Tx Broadcasted {status.txHash}</div>
     )
 
   return (