+++ packages/client/src/index.ts
@@ -3,18 +3,16 @@ export * from "./re-exports"
 export * from "./descriptors"
 export { InvalidTxError } from "./tx"
 export type {
-  TxEvent,
-  TxBroadcastEvent,
-  TxSigned,
+  Transaction,
   TxBroadcasted,
-  TxBestBlocksState,
-  TxInBestBlocksNotFound,
-  TxInBestBlocksFound,
+  TxCreated,
+  TxEntry,
+  TxEvent,
   TxEventsPayload,
   TxFinalized,
   TxFinalizedPayload,
-  Transaction,
-  TxEntry,
+  TxInBestBlock,
+  TxNotInBestBlock,
 } from "./tx"
 export type { EventPhase } from "./event"
 export type { TxCreator } from "@polkadot-api/tx-creator"